changed: visualization interface to float from signed 16bit
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/AudioEngine@32817 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/addons/include/xbmc_vis_types.h
@@ -97,7 +97,7 @@ extern "C"
   struct Visualisation
   {
     void (__cdecl* Start)(int iChannels, int iSamplesPerSec, int iBitsPerSample, const char* szSongName);
-    void (__cdecl* AudioData)(const short* pAudioData, int iAudioDataLength, float *pFreqData, int iFreqDataLength);
+    void (__cdecl* AudioData)(const float* pAudioData, int iAudioDataLength, float *pFreqData, int iFreqDataLength);
     void (__cdecl* Render) ();
     void (__cdecl* GetInfo)(VIS_INFO *info);
     bool (__cdecl* OnAction)(long flags, const void *param);
